requirements.txt
langchain-community
faiss-cpu
langchain-openai
langchain-text-splitters
langchain-voyageai
python-dotenv
beautifulsoup4
fastapi
uvicorn[standard]
pytest